user specify addresses
