+++ my-app/src/App.js
@@ -1,13 +1,11 @@
 import React from 'react';
 import './App.css';
 import Map from './Map.js';
-import Addresses from './Addresses'
 
 function App() {
 
   return (
     <div className="App">
-      <Addresses />
       <Map/>
     </div>      
   );
